backend/requirements.txt
fastapi
uvicorn
sqlalchemy
chromadb
pydantic
groq
google-generativeai
requests
pytest
